refactored structure of codebase
