+++ src/features/sponsors/data/payload/getSponsors.ts
